+++ src/main/java/mod/gottsch/forge/treasure2/core/capability/DurabilityHandler.java
@@ -20,11 +20,15 @@ package mod.gottsch.forge.treasure2.core.capability;
 import java.util.List;
 
 import mod.gottsch.forge.treasure2.Treasure;
+import mod.gottsch.forge.treasure2.core.config.Config;
+import mod.gottsch.forge.treasure2.core.event.WorldEventHandler;
+import mod.gottsch.forge.treasure2.core.util.TreasureDataFixer;
 import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
@@ -39,11 +43,23 @@ public class DurabilityHandler implements IDurabilityHandler {
 	private static final String DURABILITY_TAG = "durability";
 	private static final String INFINITE_TAG = "infinite";
 	
-	// the max value the durability can be set to
-	private int maxDurability = MAX_DURABILITY; // this is kinda useless and confusing
-	
-	// the durability of the capability
-	private int durability;
+	/*
+	 * the max value the durability can be set to.
+	 * ex. if merging two keys of 25 durability each and maxDurability = 40
+	 * then the durability of the merged key will be 40
+	 * 	ie Math.min(key1.durability + key2.durability, maxDurability)
+	 * ex. if creating a new itemstack and setting the durability, the
+	 * max it will have is maxDurability.
+	 */
+	private int maxDurability = MAX_DURABILITY;
+
+	/*
+	 * the durability of newly created item stacks.
+	 */
+	private int defaultDurability;
+
+	// the current durability of the capability.
+	private int durability = Integer.MIN_VALUE;
 	
 	// is this durability infinite (as opposed to the finite value of the durability property)
 	private boolean infinite;
@@ -58,7 +74,6 @@ public class DurabilityHandler implements IDurabilityHandler {
 	 * 
 	 */
 	public DurabilityHandler() {
-		setDurability(MAX_DURABILITY);
 	}
 	
 	public DurabilityHandler(int durability) {
@@ -75,6 +90,7 @@ public class DurabilityHandler implements IDurabilityHandler {
 	public CompoundTag save() {
 		CompoundTag mainTag = new CompoundTag();
 		try {
+			mainTag.putInt("defaultDurability", getDefaultDurability());
 			mainTag.putInt(DURABILITY_TAG, getDurability());
 			mainTag.putInt(MAX_DURABILITY_TAG, getMaxDurability());
 			mainTag.putBoolean(INFINITE_TAG, isInfinite());
@@ -93,7 +109,10 @@ public class DurabilityHandler implements IDurabilityHandler {
 			if (compound.contains(MAX_DURABILITY_TAG)) {
 				setMaxDurability(compound.getInt(MAX_DURABILITY_TAG));
 			}
-			
+
+			if (compound.contains("defaultDurability")) {
+				setDefaultDurability(compound.getInt("defaultDurability"));
+			}
 			if (compound.contains(DURABILITY_TAG)) {
 				setDurability(compound.getInt(DURABILITY_TAG));
 			}
@@ -113,13 +132,14 @@ public class DurabilityHandler implements IDurabilityHandler {
 	
 	@Override
 	public void copyTo(ItemStack stack) {
-		stack.getCapability(TreasureCapabilities.DURABILITY).ifPresent(cap -> {
+		stack.getCapability(TreasureCapabilities.DURABILITY).ifPresent(handler -> {
 			// note: set max first
-			cap.setMaxDurability(getMaxDurability());
-			cap.setDurability(getDurability());
-			cap.setInfinite(isInfinite());			
-			cap.setMaxRepairs(getMaxRepairs());
-			cap.setRepairs(getRepairs());			
+			handler.setMaxDurability(getMaxDurability());
+			((DurabilityHandler)handler).setDefaultDurability(getDefaultDurability());
+			((DurabilityHandler)handler).setDurability(getDurability());
+			handler.setInfinite(isInfinite());
+			handler.setMaxRepairs(getMaxRepairs());
+			handler.setRepairs(getRepairs());
 		});
 	}
 	
@@ -136,12 +156,38 @@ public class DurabilityHandler implements IDurabilityHandler {
 			tooltip.add(Component.translatable("tooltip.durability.amount.infinite").withStyle(ChatFormatting.WHITE));
 		}
 	}
-	
+
 	@Override
-	public int getDurability() {
+	public int getDefaultDurability() {
+		return defaultDurability;
+	}
+
+	@Override
+	public void setDefaultDurability(int defaultDurability) {
+		this.defaultDurability = defaultDurability;
+	}
+
+	@Override
+	public int durability(Item item) {
+		if (durability == Integer.MIN_VALUE) {
+			// check if config is loaded
+			// NOTE both sides are loaded from the SERVER config values.
+			if (WorldEventHandler.isServerLoaded() || WorldEventHandler.isClientLoaded()) {
+				durability = TreasureDataFixer.KEYS_CONFIG_MAP.get(item);
+				// update the default durability to the same value.
+				// NOTE default durability does/should not change after being set.
+				defaultDurability = durability;
+			} else {
+				return defaultDurability;
+			}
+		}
 		return durability;
 	}
-	
+
+	protected int getDurability() {
+		return durability;
+	}
+
 	@Override
 	public void setDurability(int maxDamage) {
 		if (maxDamage > getMaxDurability()) {