+++ src/main/java/mod/gottsch/forge/treasure2/core/event/WorldEventHandler.java
@@ -104,4 +104,12 @@ public class WorldEventHandler {
 			}
 		}
 	}
+
+	public static boolean isServerLoaded() {
+		return isLoaded;
+	}
+
+	public static boolean isClientLoaded() {
+		return isClientLoaded;
+	}
 }