+++ src/main/java/mod/gottsch/forge/treasure2/core/item/JewelledKey.java
@@ -17,8 +17,6 @@
  */
 package mod.gottsch.forge.treasure2.core.item;
 
-import java.util.List;
-
 import mod.gottsch.forge.treasure2.core.util.LangUtil;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
@@ -26,6 +24,8 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
+
+import java.util.List;
 
 /**
  * 
@@ -35,7 +35,11 @@ import net.minecraft.world.level.Level;
 public class JewelledKey extends KeyItem {
 
 	public JewelledKey(Item.Properties properties) {
-		super(properties);
+		this(properties, DEFAULT_MAX_USES);
+	}
+
+	public JewelledKey(Item.Properties properties, int durability) {
+		super(properties, durability);
 		addFitsLock(lock -> {
 			return lock.getCategory() == KeyLockCategory.GEMS;
 		});