+++ src/main/java/mod/gottsch/forge/treasure2/core/item/SkeletonKey.java
@@ -17,9 +17,6 @@
  */
 package mod.gottsch.forge.treasure2.core.item;
 
-import java.awt.TextComponent;
-import java.util.List;
-
 import mod.gottsch.forge.gottschcore.enums.IRarity;
 import mod.gottsch.forge.treasure2.core.enums.Rarity;
 import mod.gottsch.forge.treasure2.core.registry.KeyLockRegistry;
@@ -31,6 +28,8 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
 
+import java.util.List;
+
 /**
  * 
  * @author Mark Gottschling on Feb 2, 2018
@@ -38,12 +37,16 @@ import net.minecraft.world.level.Level;
  */
 public class SkeletonKey extends KeyItem {
 
+	public SkeletonKey(Item.Properties properties) {
+		this(properties, DEFAULT_MAX_USES);
+	}
+
 	/**
 	 * 
 	 * @param properties
 	 */
-	public SkeletonKey(Item.Properties properties) {
-		super(properties);
+	public SkeletonKey(Item.Properties properties, int durability) {
+		super(properties, durability);
 		
 		// add the default fitsLock predicates
 		addFitsLock(lock -> {