+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.8.2] - 2024-05-20
+
+### Changed
+
+-Fixed crafted Keys from only having 0 durability/usage on server.
+-Fixed reading datapacks in flat/exploded directory format.
+-KeyItem now takes a default durability value in the constructor.
+
 ## [3.8.1] - 2024-02-15
 
 ### Changed